Removed warning about signed - unsigned comparison

+++ include/EventParametersParser.h
@@ -137,7 +137,7 @@ private:
     if (value.size() != strlen(expected))
       return false;
 
-    for (int i = 0; i < value.size(); ++i) {
+    for (size_t i = 0; i < value.size(); ++i) {
       if (tolower(value[i]) != expected[i]) {
         return false;
       }
